// @types/api-datas.d.ts
export interface CityData {
    geonameId: string;
    nameFr: string;
    nameEn: string;
    countryId: string;
    countryCode: string;
    countryIndexFr: string;
    countryIndexEn: string;
    continentCode: string;
    continentIndexFr: string;
    continentIndexEn: string;
    timezone: string;
    latitude: string;
    longitude: string;
}